Added Orbiter's internal event : ScreenRendered Fixed few warnings.

+++ src/Orbiter/MediaBrowserMouseHandler.cpp
@@ -156,7 +156,7 @@ void MediaBrowserMouseHandler::Move(int X,int Y,int PK_Direction)
 	// issue becaues pObj_ListGrid_Active->m_FixedRowHeight doesn't have the fractional amount and can get off
 	// int Row = (Y-pObj_ListGrid_Active->m_rPosition.Y)/(pObj_ListGrid_Active->m_FixedRowHeight+1); // +1 because there's a blank pixel under each one
 	// So use this method when there's a DesignObj for the cell
-	int Row = (Y-pObj_ListGrid_Active->m_rPosition.Y)/((float) pObj_ListGrid_Active->m_rPosition.Height / pObj_ListGrid_Active->m_MaxRow);
+	int Row = int((Y-pObj_ListGrid_Active->m_rPosition.Y)/((float) pObj_ListGrid_Active->m_rPosition.Height / pObj_ListGrid_Active->m_MaxRow));
 
 #ifdef DEBUG
 	g_pPlutoLogger->Write(LV_WARNING, "MediaBrowserMouseHandler::Move(%d, %d, %d) row %d active %p=%p", X, Y, PK_Direction,Row,pObj_ListGrid_Active,m_pObj_PicGrid);
